# openapi-mvp.yaml
openapi: 3.1.0
info:
  title: Work Discovery AI MVP API
  version: 0.9.0
  description: >-
    텍스트·음성 인터뷰를 근거가 연결된 업무 모델로 변환하고,
    개선·자동화 후보, G1 솔루션 명세, CODEGEN_READY 구현 패키지를 생성하는 MVP API 계약.
servers:
  - url: https://api.example.local/v1
security:
  - bearerAuth: []
...
paths:
  /projects:
    post:
      tags: [Projects]
      operationId: createProject
      summary: 새 업무 분석 프로젝트 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/Problem'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      operationId: getProject
      summary: 프로젝트 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: 프로젝트와 파생 데이터 삭제 요청
      responses:
        '202':
          description: 삭제 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionJob'
        '404':
          $ref: '#/components/responses/Problem'
  /projects/{projectId}/interviews:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Interviews]
      operationId: createInterview
      summary: 인터뷰 세션 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInterviewRequest'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interview'
        '409':
          $ref: '#/components/responses/Problem'
  /interviews/{interviewId}:
    parameters:
      - $ref: '#/components/parameters/InterviewId'
    get:
      tags: [Interviews]
      operationId: getInterview
      summary: 인터뷰 상태 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interview'
        '404':
          $ref: '#/components/responses/Problem'
  /interviews/{interviewId}/consent:
    parameters:
      - $ref: '#/components/parameters/InterviewId'
    post:
      tags: [Interviews]
      operationId: confirmConsent
      summary: AI·데이터·음성 처리 동의 기록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
      responses:
        '200':
          description: 동의 기록됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interview'
        '422':
          $ref: '#/components/responses/Problem'
  /interviews/{interviewId}/answers:
    parameters:
      - $ref: '#/components/parameters/InterviewId'
    post:
      tags: [Interviews]
      operationId: submitTextAnswer
      summary: 텍스트 답변 제출
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
      responses:
        '202':
          description: 분석 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJob'
        '409':
          $ref: '#/components/responses/Problem'
...
  /projects/{projectId}/work-model:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [WorkModels]
      operationId: getWorkModel
      summary: 최신 업무 모델 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: work-model-v1.schema.json에 의해 검증되는 객체
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [WorkModels]
      operationId: patchWorkModel
      summary: 사용자 수정사항 반영 및 새 버전 생성
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/JsonPatchOperation'
      responses:
        '200':
          description: 새 버전
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '409':
          $ref: '#/components/responses/Problem'
  /projects/{projectId}/work-model/validate:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [WorkModels]
      operationId: validateWorkModel
      summary: 사용자가 업무 지도와 추론을 확인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateModelRequest'
      responses:
        '200':
          description: 검증 상태 갱신
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '422':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    InterviewId:
      name: interviewId
      in: path
      required: true
      schema:
        type: string
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 최신 업무 모델 버전 ETag
      schema:
        type: string
  responses:
    Problem:
      description: RFC 9457 스타일 오류 응답
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [title, purpose]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        purpose:
          type: string
          minLength: 1
          maxLength: 2000
        data_region:
          type: string
          maxLength: 50
    Project:
      type: object
      additionalProperties: false
      required: [project_id, title, purpose, status, created_at, updated_at]
      properties:
        project_id:
          type: string
        title:
          type: string
        purpose:
          type: string
        status:
          type: string
          enum: [DRAFT, INTERVIEWING, ANALYZING, BLUEPRINT_READY, PAUSED, DELETION_PENDING, DELETED]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateInterviewRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          type: string
          enum: [QUICK, STANDARD, DEEP]
        language:
          type: string
          default: ko-KR
    ConsentRequest:
      type: object
      additionalProperties: false
      required: [ai_use, data_processing, audio_recording, audio_retention]
      properties:
        ai_use:
          type: boolean
        data_processing:
          type: boolean
        audio_recording:
          type: boolean
        audio_retention:
          type: string
          enum: [DELETE_AFTER_TRANSCRIPT_CONFIRMATION, KEEP_UNTIL_PROJECT_DELETION, NOT_APPLICABLE]
    Interview:
      type: object
      additionalProperties: false
      required: [interview_id, project_id, state, mode, round, created_at, updated_at]
      properties:
        interview_id:
          type: string
        project_id:
          type: string
        state:
          type: string
        mode:
          type: string
        round:
          type: integer
          minimum: 0
        coverage:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AnswerRequest:
      type: object
      additionalProperties: false
      required: [question_id, text]
      properties:
        question_id:
          type: string
        text:
          type: string
          minLength: 1
          maxLength: 50000
        client_answer_id:
          type: string
          description: 재시도 시 중복 생성을 막기 위한 클라이언트 멱등 키
...
    ValidateModelRequest:
      type: object
      additionalProperties: false
      required: [model_version, confirmed, corrections]
      properties:
        model_version:
          type: integer
          minimum: 1
        confirmed:
          type: boolean
        corrections:
          type: array
          items:
            $ref: '#/components/schemas/JsonPatchOperation'
        notes:
          type: string
          maxLength: 4000
    ValidationResult:
      type: object
      additionalProperties: false
      required: [model_id, version, status, understanding_gate]
      properties:
        model_id:
          type: string
        version:
          type: integer
        status:
          type: string
        understanding_gate:
          type: object
          additionalProperties: true
...
    JsonPatchOperation:
      type: object
      required: [op, path]
      properties:
        op:
          type: string
          enum: [add, remove, replace, move, copy, test]
        path:
          type: string
        from:
          type: string
        value: {}
    ProcessingJob:
      type: object
      additionalProperties: false
      required: [job_id, status, submitted_at]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED]
        submitted_at:
          type: string
          format: date-time
    DeletionJob:
      type: object
      additionalProperties: false
      required: [deletion_job_id, project_id, status, requested_at]
      properties:
        deletion_job_id:
          type: string
        project_id:
          type: string
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED]
        requested_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [type, title, status, trace_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        trace_id:
          type: string